refactor: extract random utils into common
